Ensure apt cache is updated before install.
Without cache being up to date, apt is unable to fetch packages from the
internet and an error is thrown.

If, in case, the cache is updated in a separate task in another role
preceding this one; the update is not executed. There is an hour of
acceptable cache time, meaning that if within an hour, an update on
cache was made; the update is ignored for that task.

Having the cache updated, fixes problems on package installation and
ensures this role works without any problem nor any additional step if
the cache is already up to date.