Replace Ubuntu init.d script
This fixes random issues running provisioning for a Vagrant VM with Ubuntu
14.04 and mysql-community-server 5.6.25-1ubuntu12.04.

/etc/init.d/mysql greps for the mysql process to determine if mysql is
running. When running, the process count includes the mysqld process and
the grep for that process (count=2). When running under vagrant using
ansible to provision, 5% - 10% of the time the grep is not in the
process list, causing false negatives when detecting the mysqld process
is running. When provisioning, this results in refusing to stop the
service and failing to start the service.

This new init script ignores the grep process, making the run detection
more reliable.

Submitted upstream as http://bugs.mysql.com/bug.php?id=77696