Allow mysql_config_include_files to contain raw content
In some cases, specifying a known src path in `mysql_config_include_files`
is not possible / complicated and it may be much cleaner to provide raw
file content to be sent. In that case, we need to use the
`ansible.builtin.copy` module. We can extend the current implementation
by simply filtering on whether `src` or `content` is set in each item
in `mysql_config_include_files`