fix: mysql version conditionals
Use ansible `version()` filter instead of substring detection.

cf. the official documentation:
https://docs.ansible.com/ansible/latest/collections/ansible/builtin/version_test.html

This will avoid potential versions detection false positives, with for example:
```
'8.0.' in mysql_cli_version
```
That could return `true` if the version was `11.8.0-MariaDB`.

NB:
This removes `query_cache_*` variables on MariaDB which do not work the same
way as on MySQL, cf.
https://mariadb.com/kb/en/server-system-variables/#query_cache_type
```
Warning: Starting from MariaDB 10.1.7, query_cache_type is automatically set to
ON if the server is started with the query_cache_size set to a non-zero (and
non-default) value. This will happen even if query_cache_type is explicitly set
to OFF in the configuration.
```
It is misleading to have `query_cache_type: 0` by default and still enabling
the `query_cache` feature, so this commit restricts `query_cache` explicit
configuration to MySQL.
This effectively resets the query cache to its default upstream value on
MariaDB:
* query cache enabled
* query cache size: 1M
* query cache limit: 1M

Signed-off-by: Alexis Thietard <alexis.thietard@rtone.fr>