fix: query caching features on mariadb
This fixes commit e21a22afd48ce2f5e1707ed1875b33ecd162b396

The mentioned commit introduced a bug, preventing users to tune the query
caching when using MariaDB, by removing `query_cache_*` variables from being
deployed in the destination my.cnf file.

This is due to me not understanding well the MariaDB documentation, and
different behaviors of query caching features between:
* MySQL versions < 8.0 where query caching was possible
* MySQL versions ≥ 8.0 where query caching was removed
* MariaDB versions < 10.1.7 where query caching was similar to MySQL before 8.0
* MariaDB versions ≥ 10.1.7 where query caching variables behavior changed,
  depending on other variables values...

This is quite complex and it might change again in the future.

This commit adds a comment describing the behavior of the three role variables:
* `mysql_query_cache_type`
* `mysql_query_cache_size`
* `mysql_query_cache_limit`
Accross all MySQL / MariaDB versions in order to clarify all of this for the
end user that might need this.

This commit fixes the bug by reintroducing these variables into the my.cnf
destination file.

Signed-off-by: Alexis Thietard <alexis.thietard@rtone.fr>